--- frontend/src/components/ChatSession.js
@@ -1,26 +1,0 @@
-import React from 'react';
-    import Proptypes from 'prop-types';
-    import { format } from 'date-fns';
-
-    const ChatSession = props => {
-      const { messages } = props;
-      return messages.map(message => {
-        const time = format(new Date(`${message.updatedAt}`), 'HH:mm');
-
-        return (
-          <li className="message" key={message.id}>
-            <div>
-              <span className="user-id">{message.senderId}</span>
-              <span>{message.text}</span>
-            </div>
-            <span className="message-time">{time}</span>
-          </li>
-        );
-      });
-    };
-
-    ChatSession.propTypes = {
-      messages: Proptypes.arrayOf(Proptypes.object).isRequired,
-    };
-
-    export default ChatSession;
